refactor: move all type exports to 'src/types/index.ts'
