+++ src/plugins/quarterOfYear/index.ts
@@ -1,6 +1,12 @@
-import type { EsDay, EsDayFactory, EsDayPlugin, FormattingTokenDefinitions, UnitType } from 'esday'
+/**
+ * quarterOfYear plugin
+ *
+ * This plugin adds 'quarter' and the formatting token 'Q' to EsDay.
+ */
+
+import type { EsDay, EsDayPlugin, FormattingTokenDefinitions } from 'esday'
 import { C, prettyUnit } from '~/common'
-import type { UnitIsoWeek } from '~/common'
+import type { UnitIsoWeek, UnitType } from '~/types'
 
 declare module 'esday' {
   interface EsDay {
@@ -9,7 +15,7 @@ declare module 'esday' {
   }
 }
 
-const quarterOfYearPlugin: EsDayPlugin<{}> = (_, dayClass, dayFactory: EsDayFactory) => {
+const quarterOfYearPlugin: EsDayPlugin<{}> = (_, dayClass, dayFactory) => {
   const proto = dayClass.prototype
 
   // @ts-expect-error function is compatible with its overload
